Cleaned Instance Property Sheet
The layout of type interface sheet is cleaned and the untyped subapp
sheet and the instance property sheet are better seperated.
