+++ plugins/org.eclipse.fordiac.ide.application/src/org/eclipse/fordiac/ide/application/properties/AbstractInstanceSection.java
@@ -1,0 +1,145 @@
+/*******************************************************************************
+ * Copyright (c) 2022 Primetals Technologies Austria GmbH, Martin Erich Jobst
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   Alois Zoitl - Extrated and refined from InstancePropertySection
+ *******************************************************************************/
+package org.eclipse.fordiac.ide.application.properties;
+
+import org.eclipse.fordiac.ide.application.commands.ResizeGroupOrSubappCommand;
+import org.eclipse.fordiac.ide.application.editparts.SubAppForFBNetworkEditPart;
+import org.eclipse.fordiac.ide.gef.properties.AbstractSection;
+import org.eclipse.fordiac.ide.model.commands.change.ChangeCommentCommand;
+import org.eclipse.fordiac.ide.model.commands.change.ChangeNameCommand;
+import org.eclipse.fordiac.ide.model.libraryElement.BlockFBNetworkElement;
+import org.eclipse.fordiac.ide.model.libraryElement.SubApp;
+import org.eclipse.fordiac.ide.ui.FordiacMessages;
+import org.eclipse.fordiac.ide.ui.editors.EditorUtils;
+import org.eclipse.gef.EditPart;
+import org.eclipse.gef.commands.Command;
+import org.eclipse.jface.action.IAction;
+import org.eclipse.jface.layout.GridDataFactory;
+import org.eclipse.swt.SWT;
+import org.eclipse.swt.layout.GridData;
+import org.eclipse.swt.layout.GridLayout;
+import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Label;
+import org.eclipse.swt.widgets.Text;
+import org.eclipse.ui.IActionBars;
+import org.eclipse.ui.actions.ActionFactory;
+import org.eclipse.ui.views.properties.tabbed.TabbedPropertySheetPage;
+
+abstract class AbstractInstanceSection extends AbstractSection {
+
+	private Text nameText;
+	private Text commentText;
+
+	IAction[] defaultCopyPasteCut = new IAction[3];
+	private TabbedPropertySheetPage tabbedPropertySheetPage;
+
+	@Override
+	public void createControls(final Composite parent, final TabbedPropertySheetPage tabbedPropertySheetPage) {
+		super.createControls(parent, tabbedPropertySheetPage);
+		parent.setLayout(new GridLayout(1, false));
+		GridDataFactory.fillDefaults().align(GridData.FILL, GridData.FILL).grab(true, true).applyTo(parent);
+	}
+
+	@Override
+	protected void performRefresh() {
+		if ((getType() != null) && (!nameText.isDisposed() && !nameText.getParent().isDisposed())) {
+			nameText.setText(getType().getName() != null ? getType().getName() : ""); //$NON-NLS-1$
+			commentText.setText(getType().getComment() != null ? getType().getComment() : ""); //$NON-NLS-1$
+		}
+	}
+
+	@Override
+	protected BlockFBNetworkElement getType() {
+		if (type instanceof final BlockFBNetworkElement fbNetworkElement) {
+			return fbNetworkElement;
+		}
+		return null;
+	}
+
+	@Override
+	public void aboutToBeShown() {
+		// this can be removed once copy/paste for old tables is no longer used
+		final IActionBars bars = getActionBars();
+		if (bars != null) {
+			defaultCopyPasteCut[0] = bars.getGlobalActionHandler(ActionFactory.COPY.getId());
+			bars.setGlobalActionHandler(ActionFactory.COPY.getId(), null);
+			defaultCopyPasteCut[1] = bars.getGlobalActionHandler(ActionFactory.PASTE.getId());
+			bars.setGlobalActionHandler(ActionFactory.PASTE.getId(), null);
+			defaultCopyPasteCut[2] = bars.getGlobalActionHandler(ActionFactory.CUT.getId());
+			bars.setGlobalActionHandler(ActionFactory.CUT.getId(), null);
+			bars.updateActionBars();
+		}
+
+		super.aboutToBeShown();
+	}
+
+	@Override
+	public void aboutToBeHidden() {
+		// this can be removed once copy/paste for old tables is no longer used
+		final IActionBars bars = getActionBars();
+		if (bars != null) {
+			bars.setGlobalActionHandler(ActionFactory.COPY.getId(), defaultCopyPasteCut[0]);
+			bars.setGlobalActionHandler(ActionFactory.PASTE.getId(), defaultCopyPasteCut[1]);
+			bars.setGlobalActionHandler(ActionFactory.CUT.getId(), defaultCopyPasteCut[2]);
+			bars.updateActionBars();
+		}
+
+		super.aboutToBeHidden();
+	}
+
+	protected IActionBars getActionBars() {
+		if (tabbedPropertySheetPage != null && tabbedPropertySheetPage.getSite() != null) {
+			return tabbedPropertySheetPage.getSite().getActionBars();
+		}
+		return null;
+	}
+
+	protected void createNameInput(final Composite parent) {
+		getWidgetFactory().createLabel(parent, FordiacMessages.Name + ":"); //$NON-NLS-1$
+		nameText = createGroupText(parent, true);
+		nameText.addModifyListener(e -> {
+			removeContentAdapter();
+			executeCommand(ChangeNameCommand.forName(getType(), nameText.getText()));
+			addContentAdapter();
+		});
+	}
+
+	protected void createCommentInput(final Composite parent) {
+		final Label commentLabel = getWidgetFactory().createLabel(parent, FordiacMessages.Comment + ":"); //$NON-NLS-1$
+		GridDataFactory.fillDefaults().align(SWT.LEFT, SWT.TOP).grab(false, false).applyTo(commentLabel);
+
+		commentText = createGroupText(parent, true, SWT.BORDER | SWT.MULTI | SWT.V_SCROLL);
+		GridDataFactory.fillDefaults().align(SWT.FILL, SWT.TOP).grab(true, false)
+				.hint(SWT.DEFAULT, 3 * commentText.getLineHeight()).applyTo(commentText);
+		commentText.addModifyListener(e -> {
+			removeContentAdapter();
+			final Command cmd = createChangeCommentCommand();
+			executeCommand(cmd);
+			addContentAdapter();
+		});
+	}
+
+	protected Command createChangeCommentCommand() {
+		Command cmd = new ChangeCommentCommand(getType(), commentText.getText());
+		if (EditorUtils.getGraphicalViewerFromCurrentActiveEditor() != null && getType() instanceof SubApp) {
+			final EditPart editPart = EditorUtils.getGraphicalViewerFromCurrentActiveEditor()
+					.getEditPartForModel(getType());
+			if (editPart instanceof final SubAppForFBNetworkEditPart subAppforFBNetworkEditPart
+					&& subAppforFBNetworkEditPart.getContentEP() != null) {
+				cmd = cmd.chain(new ResizeGroupOrSubappCommand(subAppforFBNetworkEditPart.getContentEP()));
+			}
+		}
+		return cmd;
+	}
+
+}
